# Speed up initialization of Longitude