--- a/CHANGES.rst
+++ b/CHANGES.rst
@@ -505,10 +505,8 @@ astropy.coordinates
 
 - Refactored ``SkyCoord`` initializer to improve performance and code clarity.
   [#7958]
-astropy.coordinates
-^^^^^^^^^^^^^^^^^^^
 
-- Sped up initialization of Longitude by ~40%. [#7616]
+- Sped up initialization of ``Longitude`` by ~40%. [#7616]
 
 astropy.units
 ^^^^^^^^^^^^^