This just avoids an extra ``Angle`` initialization if ``wrap_angle`` is already an ``Angle`` object.

### Before

```python
In [4]: ra = 3 * u.deg

In [5]: %timeit Longitude(ra)
71.3 µs ± 4.75 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

### After

```python
In [6]: ra = 3 * u.deg

In [7]: %timeit Longitude(ra)
51.2 µs ± 2.8 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)
```

This also includes a speed up of ``Unit.to`` when the unit passed is the same as the original unit. This can actually happen quite commonly in code where in most cases the default will be for the unit to be the same (such as in Longitude):

### Before

```python
In [6]: %timeit u.deg.to(u.deg)
1.68 µs ± 38.5 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

### After

```python
In [10]: %timeit u.deg.to(u.deg)
279 ns ± 23.8 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```