@mhvk 2018-07-01
@astrofrog - could you split the two? I ask in part since the speed-up to `Unit.to` is, I think, better done further up the chain, in `Quantity`. There, the check on units being identical is done for `to_value`, but, contrary to what I thought, not for `to`. The latter may be because a copy is expected, which made me think it didn't matter (wrongly).  But on copies: your change means that `Unit.to` no longer is guaranteed to return a copy, which I think is risky.