@astrofrog 2018-07-12
@mhvk - just a question about your comment:

> I ask in part since the speed-up to Unit.to is, I think, better done further up the chain, in Quantity

The issue is that not everyone uses Quantity - in fact the code I was trying to optimize specifically uses just e.g. ``factor = unit.to(other_unit)``. So I do think we want to optimize it in Unit.to as well. I'm fine with returning a copy of the value though.