@mhvk 2018-07-12
> I'm fine with returning a copy of the value though.
OK. Still, let's do it elsewhere. E.g., you cannot be sure that `value` is an array, so need to use `copy.copy(value)` or something like it. We may also want to consider whether `Unit.to` shoudn't have an `inplace` flag...