@astrofrog 2018-07-12
@mhvk - I think we can still speed things up for the case of the default value=1.0 though, so something like:

```python
def to(self, other, value=1.0, equivalencies=[]):
    if other is self and value == 1.0:
            return value
        else:
            return self._get_converter(other, equivalencies=equivalencies)(value)
```

Would you be happy for something like this for now?