@mhvk 2018-07-12
That would break if `value` were an array...  If that's the one you want, and you do not care about equivalencies, you probably want to use `unit._to(other_unit)` - it is used quite frequently, so perhaps we should expose that? (But let's discuss in a separate PR, so we can get the `Longitude` fix in!)