@astrofrog 2018-07-12
@mhvk - I simplified this PR. I'm still not super happy with the fact that in the default case where no wrap angle is specified, we are basically doing ``self._wrap_angle = Angle(Angle(360 * u.deg), copy=False)``, which seems unnecessary. One option would be to do:

```
        if value is self._default_wrap_angle:
            self._wrap_angle = value
        else:
            self._wrap_angle = Angle(value, copy=False)
```

What do you think?