@mhvk 2018-07-12
I like the idea, but think one might want to change the initializer instead; maybe something like:
```
    _wrap_angle = _default_wrap_angle = Angle(360., u.deg)
    _default_wrap_angle.flags.writable = False

    def __new__(...):
...
        if wrap_angle is None:
            wrap_angle = getattr(angle, 'wrap_angle', None)
        if wrap_angle is not None:
            self.wrap_angle = wrap_angle
```
Note the setting to read-only: one risk with not making copies is that one can accidentally change the default and then make the whole class unusable...