@astrofrog 2018-07-12
@mhvk - aren't scalar quantities immutable?