@mhvk 2018-07-12
Perhaps scalar quantities should be immutable, but at present they aren't (they're just array scalar, which are also not immutable).