@astrofrog 2018-07-13
@mhvk - just out of curiosity, how would you modify the value of a scalar quantity in-place?