@mhvk 2018-07-13
```
In [4]: Longitude._default_wrap_angle[()] = 0.

In [5]: Longitude._default_wrap_angle
Out[5]: <Angle 0. deg>
```