@adrn 2018-10-26
I think @mhvk's solution doesn't help because we'd still have to call `self._wrap_internal()` after the code he suggested, and it looks like that's where one of the bottlenecks is.