@adrn 2018-11-02
@mhvk I was referring to the code you suggested [in this comment](https://github.com/astropy/astropy/pull/7616#issuecomment-404590118) - I agree that this PR is straightforward and the speedup helps for scalars.